http-util: extract handle_prometheus handler
This pulls the logic out of materialized into a new crate so other
binaries can depend on it. It also wires the handler up to the existing
http console servers in storaged and computed.
